refactor: split file upload page
